refactor(deepagents): move HITL glue out of FilesystemMiddleware
`FilesystemMiddleware` itself doesn't use `_build_interrupt_on_from_permissions`
or the `when` predicates — only `deepagents.graph` does, to wire
`HumanInTheLoopMiddleware`. The HITL bridge sat in `filesystem.py` only
because the path-arg-per-tool table lived alongside the tool definitions.

Move that glue into a new `deepagents/middleware/_fs_interrupt.py`:
`ToolScope`, `_FS_TOOL_PATH_ARGS`, the predicate constructors, and
`_build_interrupt_on_from_permissions`. `graph.py` now imports the
builder from the new module. `filesystem.py` drops 168 lines and the
now-unused `InterruptOnConfig`/`to_posix_path` imports — keeps
`FilesystemPermission`, `_check_fs_permission`, filter helpers, tools.

Also relocate the generic path utilities `_glob_anchor`/`_paths_overlap`
to `deepagents/backends/utils.py` alongside `to_posix_path` and
`validate_path`. They're not filesystem-middleware-specific.

Test imports updated; no behavior change. 1652 tests pass; `make lint`
clean.
